refactor: Controller / Domain / Util 분리
