+++ src/main/java/ladder/domain/Point.java
@@ -1,6 +1,5 @@
-package ladder;
+package ladder.domain;
 
-import java.util.List;
 import java.util.Objects;
 
 public class Point {
