+++ src/test/java/ladder/UserTest.java
@@ -1,5 +1,6 @@
 package ladder;
 
+import ladder.domain.User;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
